feat: rename, become a preset
This changes the package name, so that we may introduce a preset for conventional-changelog to
support our non-standard commit parsing. This means that all consumers must use angular message
formats, and is an intentional design choice. Further, the existing tagger was renamed, and is still
expected to be used alongside the preset. While this makes configuration a bit messy, it is,
unfortunately, required if we're to remain a third-party extension to standard-version.

BREAKING CHANGE: Consumer expectations changed - use the preset in conjunction with the postcommit
tagger.

fix #1, re #2